Moved all LookAndFeel classes to appropriate directory. Updated project file
