+++ Builds/VisualStudio2012/open-ephys.vcxproj
@@ -321,13 +321,14 @@
     <ClCompile Include="..\..\Source\Processors\Visualization\DataWindow.cpp"/>
     <ClCompile Include="..\..\Source\Processors\Visualization\SpikeObject.cpp"/>
     <ClCompile Include="..\..\Source\Processors\Visualization\MatlabLikePlot.cpp"/>
+    <ClCompile Include="..\..\Source\UI\LookAndFeel\MaterialSliderLookAndFeel.cpp"/>
+    <ClCompile Include="..\..\Source\UI\LookAndFeel\CustomLookAndFeel.cpp"/>
     <ClCompile Include="..\..\Source\UI\CustomArrowButton.cpp"/>
     <ClCompile Include="..\..\Source\UI\GraphViewer.cpp"/>
     <ClCompile Include="..\..\Source\UI\EditorViewportButtons.cpp"/>
     <ClCompile Include="..\..\Source\UI\SignalChainManager.cpp"/>
     <ClCompile Include="..\..\Source\UI\EditorViewport.cpp"/>
     <ClCompile Include="..\..\Source\UI\ProcessorList.cpp"/>
-    <ClCompile Include="..\..\Source\UI\CustomLookAndFeel.cpp"/>
     <ClCompile Include="..\..\Source\UI\InfoLabel.cpp"/>
     <ClCompile Include="..\..\Source\UI\DataViewport.cpp"/>
     <ClCompile Include="..\..\Source\UI\ControlPanel.cpp"/>
@@ -1532,13 +1533,14 @@
     <ClInclude Include="..\..\Source\Processors\Visualization\SpikeObject.h"/>
     <ClInclude Include="..\..\Source\Processors\Visualization\Visualizer.h"/>
     <ClInclude Include="..\..\Source\Processors\Visualization\MatlabLikePlot.h"/>
+    <ClInclude Include="..\..\Source\UI\LookAndFeel\MaterialSliderLookAndFeel.h"/>
+    <ClInclude Include="..\..\Source\UI\LookAndFeel\CustomLookAndFeel.h"/>
     <ClInclude Include="..\..\Source\UI\CustomArrowButton.h"/>
     <ClInclude Include="..\..\Source\UI\GraphViewer.h"/>
     <ClInclude Include="..\..\Source\UI\EditorViewportButtons.h"/>
     <ClInclude Include="..\..\Source\UI\SignalChainManager.h"/>
     <ClInclude Include="..\..\Source\UI\EditorViewport.h"/>
     <ClInclude Include="..\..\Source\UI\ProcessorList.h"/>
-    <ClInclude Include="..\..\Source\UI\CustomLookAndFeel.h"/>
     <ClInclude Include="..\..\Source\UI\InfoLabel.h"/>
     <ClInclude Include="..\..\Source\UI\DataViewport.h"/>
     <ClInclude Include="..\..\Source\UI\ControlPanel.h"/>
@@ -2066,7 +2068,6 @@
     <None Include="..\..\JuceLibraryCode\modules\juce_audio_formats\juce_module_info"/>
     <None Include="..\..\JuceLibraryCode\modules\juce_audio_processors\juce_module_info"/>
     <None Include="..\..\JuceLibraryCode\modules\juce_audio_utils\juce_module_info"/>
-    <None Include="..\..\JuceLibraryCode\modules\juce_core\system\juce_StandardHeader.h~"/>
     <None Include="..\..\JuceLibraryCode\modules\juce_core\juce_module_info"/>
     <None Include="..\..\JuceLibraryCode\modules\juce_cryptography\juce_module_info"/>
     <None Include="..\..\JuceLibraryCode\modules\juce_data_structures\juce_module_info"/>
